fix actions
